Refactory tests
