+++ data-index/data-index-common/pom.xml
@@ -45,6 +45,10 @@
       <groupId>org.kie.kogito</groupId>
       <artifactId>kogito-events-core</artifactId>
     </dependency>
+     <dependency>
+      <groupId>io.quarkus</groupId>
+      <artifactId>quarkus-vertx-graphql</artifactId>
+    </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-jackson</artifactId>
