+++ data-index/data-index-graphql/pom.xml
@@ -43,10 +43,6 @@
     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
-      <artifactId>quarkus-vertx-graphql</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>io.quarkus</groupId>
       <artifactId>quarkus-reactive-routes</artifactId>
     </dependency>
     <dependency>
