+++ data-index/data-index-storage/data-index-storage-jpa/src/test/java/org/kie/kogito/index/jpa/query/AbstractUserTaskInstanceEntityQueryIT.java
@@ -16,15 +16,25 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.index.postgresql.storage;
+package org.kie.kogito.index.jpa.query;
 
-import org.kie.kogito.index.jpa.storage.AbstractProcessInstanceStorageIT;
-import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
+import org.kie.kogito.index.jpa.storage.UserTaskInstanceEntityStorage;
+import org.kie.kogito.index.test.query.AbstractUserTaskInstanceQueryIT;
 
-import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.junit.QuarkusTest;
+import jakarta.inject.Inject;
 
-@QuarkusTest
-@QuarkusTestResource(PostgreSqlQuarkusTestResource.class)
-public class ProcessInstanceStorageIT extends AbstractProcessInstanceStorageIT {
+public abstract class AbstractUserTaskInstanceEntityQueryIT extends AbstractUserTaskInstanceQueryIT {
+
+    @Inject
+    UserTaskInstanceEntityStorage storage;
+
+    @Override
+    public UserTaskInstanceEntityStorage getStorage() {
+        return storage;
+    }
+
+    @Override
+    protected Boolean isDateTimeAsLong() {
+        return false;
+    }
 }
