+++ data-index/data-index-storage/data-index-storage-oracle/pom.xml
@@ -40,7 +40,7 @@
     <dependency>
       <groupId>org.kie.kogito</groupId>
       <artifactId>data-index-storage-jpa</artifactId>
-    </dependency>
+     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-jdbc-oracle</artifactId>
@@ -50,8 +50,16 @@
       <artifactId>data-index-storage-api</artifactId>
       <type>test-jar</type>
       <scope>test</scope>
+      <version>${project.version}</version>
     </dependency>
     <dependency>
+       <groupId>org.kie.kogito</groupId>
+       <artifactId>data-index-storage-jpa</artifactId>
+       <type>test-jar</type>
+       <scope>test</scope>
+       <version>${project.version}</version>
+     </dependency>
+     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-junit5-mockito</artifactId>
       <scope>test</scope>
