+++ data-index/data-index-storage/data-index-storage-oracle/src/test/java/org/kie/kogito/index/oracle/query/JobEntityQueryIT.java
@@ -18,27 +18,14 @@
  */
 package org.kie.kogito.index.oracle.query;
 
-import org.kie.kogito.index.jpa.storage.JobEntityStorage;
-import org.kie.kogito.index.model.Job;
-import org.kie.kogito.index.test.query.AbstractJobQueryIT;
-import org.kie.kogito.persistence.api.Storage;
+import org.kie.kogito.index.jpa.query.AbstractJobEntityQueryIT;
 import org.kie.kogito.testcontainers.quarkus.OracleSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import jakarta.inject.Inject;
-
 @QuarkusTest
 @QuarkusTestResource(OracleSqlQuarkusTestResource.class)
-class JobEntityQueryIT extends AbstractJobQueryIT {
-
-    @Inject
-    JobEntityStorage storage;
-
-    @Override
-    public Storage<String, Job> getStorage() {
-        return storage;
-    }
+class JobEntityQueryIT extends AbstractJobEntityQueryIT {
 
 }
