+++ data-index/data-index-storage/data-index-storage-oracle/src/test/java/org/kie/kogito/index/oracle/query/UserTaskInstanceEntityQueryIT.java
@@ -18,29 +18,14 @@
  */
 package org.kie.kogito.index.oracle.query;
 
-import org.kie.kogito.index.jpa.storage.UserTaskInstanceEntityStorage;
-import org.kie.kogito.index.test.query.AbstractUserTaskInstanceQueryIT;
+import org.kie.kogito.index.jpa.query.AbstractUserTaskInstanceEntityQueryIT;
 import org.kie.kogito.testcontainers.quarkus.OracleSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import jakarta.inject.Inject;
-
 @QuarkusTest
 @QuarkusTestResource(OracleSqlQuarkusTestResource.class)
-class UserTaskInstanceEntityQueryIT extends AbstractUserTaskInstanceQueryIT {
+class UserTaskInstanceEntityQueryIT extends AbstractUserTaskInstanceEntityQueryIT {
 
-    @Inject
-    UserTaskInstanceEntityStorage storage;
-
-    @Override
-    public UserTaskInstanceEntityStorage getStorage() {
-        return storage;
-    }
-
-    @Override
-    protected Boolean isDateTimeAsLong() {
-        return false;
-    }
 }
