+++ data-index/data-index-storage/data-index-storage-postgresql/pom.xml
@@ -51,6 +51,13 @@
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
+     <dependency>
+       <groupId>org.kie.kogito</groupId>
+       <artifactId>data-index-storage-jpa</artifactId>
+       <type>test-jar</type>
+       <scope>test</scope>
+       <version>${project.version}</version>
+     </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-junit5-mockito</artifactId>
