+++ data-index/data-index-storage/data-index-storage-postgresql/src/test/java/org/kie/kogito/index/postgresql/query/JobEntityQueryIT.java
@@ -18,27 +18,13 @@
  */
 package org.kie.kogito.index.postgresql.query;
 
-import org.kie.kogito.index.jpa.storage.JobEntityStorage;
-import org.kie.kogito.index.model.Job;
-import org.kie.kogito.index.test.query.AbstractJobQueryIT;
-import org.kie.kogito.persistence.api.Storage;
+import org.kie.kogito.index.jpa.query.AbstractJobEntityQueryIT;
 import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import jakarta.inject.Inject;
-
 @QuarkusTest
 @QuarkusTestResource(PostgreSqlQuarkusTestResource.class)
-class JobEntityQueryIT extends AbstractJobQueryIT {
-
-    @Inject
-    JobEntityStorage storage;
-
-    @Override
-    public Storage<String, Job> getStorage() {
-        return storage;
-    }
-
+class JobEntityQueryIT extends AbstractJobEntityQueryIT {
 }
