+++ data-index/data-index-storage/data-index-storage-postgresql/src/test/java/org/kie/kogito/index/postgresql/query/ProcessDefinitionEntityQueryIT.java
@@ -18,31 +18,13 @@
  */
 package org.kie.kogito.index.postgresql.query;
 
-import org.kie.kogito.index.jpa.storage.ProcessDefinitionEntityStorage;
-import org.kie.kogito.index.model.ProcessDefinition;
-import org.kie.kogito.index.test.query.AbstractProcessDefinitionQueryIT;
-import org.kie.kogito.persistence.api.Storage;
+import org.kie.kogito.index.jpa.query.AbstractProcessDefinitionEntityQueryIT;
 import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import jakarta.inject.Inject;
-
 @QuarkusTest
 @QuarkusTestResource(PostgreSqlQuarkusTestResource.class)
-class ProcessDefinitionEntityQueryIT extends AbstractProcessDefinitionQueryIT {
-
-    @Inject
-    ProcessDefinitionEntityStorage storage;
-
-    @Override
-    public Storage<String, ProcessDefinition> getStorage() {
-        return storage;
-    }
-
-    @Override
-    protected Boolean isDateTimeAsLong() {
-        return false;
-    }
+class ProcessDefinitionEntityQueryIT extends AbstractProcessDefinitionEntityQueryIT {
 }
