+++ data-index/data-index-storage/data-index-storage-postgresql/src/test/java/org/kie/kogito/index/postgresql/storage/JobStorageIT.java
@@ -18,58 +18,13 @@
  */
 package org.kie.kogito.index.postgresql.storage;
 
-import java.util.UUID;
-
-import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.jupiter.api.Test;
-import org.kie.kogito.index.jpa.model.JobEntity;
-import org.kie.kogito.index.jpa.model.JobEntityRepository;
-import org.kie.kogito.index.model.Job;
-import org.kie.kogito.index.test.TestUtils;
-import org.kie.kogito.persistence.api.StorageService;
+import org.kie.kogito.index.jpa.storage.AbstractJobStorageIT;
 import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 
-import jakarta.inject.Inject;
-
 @QuarkusTest
 @QuarkusTestResource(PostgreSqlQuarkusTestResource.class)
-public class JobStorageIT extends AbstractStorageIT<JobEntity, Job> {
-
-    @Inject
-    JobEntityRepository repository;
-
-    @Inject
-    StorageService storage;
-
-    public JobStorageIT() {
-        super(Job.class);
-    }
-
-    @Override
-    public JobEntityRepository getRepository() {
-        return repository;
-    }
-
-    @Override
-    public StorageService getStorage() {
-        return storage;
-    }
-
-    @Test
-    public void testJobEntity() {
-        String jobId = UUID.randomUUID().toString();
-        String processInstanceId = UUID.randomUUID().toString();
-
-        Job job1 = TestUtils
-                .createJob(jobId, processInstanceId, RandomStringUtils.randomAlphabetic(5), UUID.randomUUID().toString(),
-                        RandomStringUtils.randomAlphabetic(10), "EXPECTED", 0L);
-        Job job2 = TestUtils
-                .createJob(jobId, processInstanceId, RandomStringUtils.randomAlphabetic(5), UUID.randomUUID().toString(),
-                        RandomStringUtils.randomAlphabetic(10), "SCHEDULED", 1000L);
-        testStorage(jobId, job1, job2);
-    }
-
+public class JobStorageIT extends AbstractJobStorageIT {
 }
